Seperate footer and header into own files
